Module Data_race_checker.Type_alias_liveness

val type_alias_liveness_block_expr : Ast.Ast_types.Var_name.t -> Ast.Ast_types.Var_name.t list -> (Ast.Ast_types.capability list -> Ast.Ast_types.capability -> bool) -> Ast.Ast_types.Var_name.t list -> Desugaring.Desugared_ast.block_expr -> Desugaring.Desugared_ast.block_expr * Ast.Ast_types.Var_name.t list

Takes in as args the Var_name.t name of the aliased object, the list of possible aliases, a capability filter function that filters out capabilities needing linear modes, a list of the current live aliases and the block expression.

Returns the updated block expression along with the live variables at the start of the block.