Module Data_race_checker.Type_concurrent_capability_access

val can_concurrently_access_capabilities : Ast.Ast_types.Class_name.t -> Desugaring.Desugared_ast.class_defn list -> Ast.Ast_types.capability -> Ast.Ast_types.capability -> bool

Checks if the two capabilities of the object of the given class can be used concurrently

val type_concurrent_capability_constraints_vars : Desugaring.Desugared_ast.class_defn list -> Desugaring.Desugared_ast.obj_var_and_capabilities list -> Ast.Ast_types.loc -> unit Core.Or_error.t

This function takes a list of free variables and types concurrent accesses of variables (it is expected that each occurrence of a given free variable corresponds to a separate thread).