Module Data_race_checker.Update_identifier_regions

val find_aliases_in_block_expr : Ast.Ast_types.Var_name.t -> Ast.Ast_types.Var_name.t list -> Desugaring.Desugared_ast.block_expr -> Ast.Ast_types.Var_name.t list

val update_identifier_regions : Ast.Ast_types.Var_name.t -> (Ast.Ast_types.region list -> Ast.Ast_types.region -> bool) -> Desugaring.Desugared_ast.identifier -> Desugaring.Desugared_ast.identifier

val update_identifier_regions_expr : Ast.Ast_types.Var_name.t -> (Ast.Ast_types.region list -> Ast.Ast_types.region -> bool) -> Desugaring.Desugared_ast.expr -> Desugaring.Desugared_ast.expr

Update all identifiers that match the given name,filtering by region list -> region -> bool - a function that given a list of all the regions, decides whether a region should be in the identifier. We also propagate the update to any aliases made of the identifier. (note this is not a symmetric propagation)

val update_identifier_regions_block_expr : Ast.Ast_types.Var_name.t -> (Ast.Ast_types.region list -> Ast.Ast_types.region -> bool) -> Desugaring.Desugared_ast.block_expr -> Desugaring.Desugared_ast.block_expr