Module Data_race_checker__Collate_region_accesses

This module collates all region accesses for objects in a given thread, and, if an identifier has multiple region options, we choose which region to use and update the identifier accordingly.

val collate_region_accesses_expr : Desugaring.Desugared_ast.class_defn list -> Desugaring.Desugared_ast.function_defn list -> Desugaring.Desugared_ast.expr -> Desugaring.Desugared_ast.expr * Desugaring.Desugared_ast.obj_var_and_regions list

val collate_region_accesses_block_expr : Desugaring.Desugared_ast.class_defn list -> Desugaring.Desugared_ast.function_defn list -> Desugaring.Desugared_ast.block_expr -> Desugaring.Desugared_ast.block_expr * Desugaring.Desugared_ast.obj_var_and_regions list