Module Data_race_checker__Data_race_checker_env

This module contains helper functions used in the data-race checking stage

val elem_in_list : 'a -> 'a list -> bool

val is_subset_of : 'a list -> 'a list -> bool

val var_lists_are_equal : Ast.Ast_types.Var_name.t list -> Ast.Ast_types.Var_name.t list -> bool

val identifier_matches_var_name : Ast.Ast_types.Var_name.t -> Desugaring.Desugared_ast.identifier -> bool

val class_has_capability : Ast.Ast_types.Class_name.t -> Ast.Ast_types.capability -> Desugaring.Desugared_ast.class_defn list -> bool

val type_has_capability : Ast.Ast_types.type_expr -> Ast.Ast_types.capability -> Desugaring.Desugared_ast.class_defn list -> bool

val identifier_has_capability : Desugaring.Desugared_ast.identifier -> Ast.Ast_types.capability -> Desugaring.Desugared_ast.class_defn list -> bool

val region_fields_have_capability : Ast.Ast_types.region -> Ast.Ast_types.Class_name.t -> Ast.Ast_types.capability -> Desugaring.Desugared_ast.class_defn list -> bool

val get_class_regions : Ast.Ast_types.Class_name.t -> Desugaring.Desugared_ast.class_defn list -> Ast.Ast_types.region list

val get_class_field_regions : Ast.Ast_types.Class_name.t -> Ast.Ast_types.Field_name.t -> Desugaring.Desugared_ast.class_defn list -> Ast.Ast_types.region list

val get_class_region_fields : Ast.Ast_types.Class_name.t -> Ast.Ast_types.Region_name.t -> Desugaring.Desugared_ast.class_defn list -> Ast.Ast_types.field_defn list

val get_identifier_name : Desugaring.Desugared_ast.identifier -> Ast.Ast_types.Var_name.t

val get_identifier_regions : Desugaring.Desugared_ast.identifier -> Ast.Ast_types.region list

val set_identifier_regions : Desugaring.Desugared_ast.identifier -> Ast.Ast_types.region list -> Desugaring.Desugared_ast.identifier

val get_function_params : Ast.Ast_types.Function_name.t -> Desugaring.Desugared_ast.function_defn list -> Ast.Ast_types.param list

val get_method_params : Ast.Ast_types.Class_name.t -> Ast.Ast_types.Method_name.t -> Desugaring.Desugared_ast.class_defn list -> Ast.Ast_types.param list

val param_to_obj_var_and_regions : Desugaring.Desugared_ast.class_defn list -> Ast.Ast_types.param -> Desugaring.Desugared_ast.obj_var_and_regions option

val params_to_obj_vars_and_regions : Desugaring.Desugared_ast.class_defn list -> Ast.Ast_types.param list -> Desugaring.Desugared_ast.obj_var_and_regions list

Convert a parameter to a representation which contains the regions it is allowed to access.

val get_method_effect_regions : Ast.Ast_types.Class_name.t -> Ast.Ast_types.Method_name.t -> Desugaring.Desugared_ast.class_defn list -> Ast.Ast_types.region list

val can_concurrently_access_regions : Ast.Ast_types.Class_name.t -> Desugaring.Desugared_ast.class_defn list -> Ast.Ast_types.region -> Ast.Ast_types.region -> bool

val reduce_expr_to_obj_id : Desugaring.Desugared_ast.expr -> Desugaring.Desugared_ast.identifier list

Use abstract interpretation to return the potential object identifiers this expression reduces to, if it reduces to an identifier.

val reduce_block_expr_to_obj_id : Desugaring.Desugared_ast.block_expr -> Desugaring.Desugared_ast.identifier list