Module Data_race_checker__Type_alias_liveness

This module checks the liveness of aliases for a object with linear capability , and updates updates the linear object reference to no longer have the linear capability when aliases are live.

(Same concept as Non-Lexical-Lifetimes in Rust)

val type_alias_liveness_block_expr : Ast.Ast_types.Var_name.t -> Ast.Ast_types.Var_name.t list -> (Ast.Ast_types.region list -> Ast.Ast_types.region -> bool) -> Ast.Ast_types.Var_name.t list -> Desugaring.Desugared_ast.block_expr -> Desugaring.Desugared_ast.block_expr * Ast.Ast_types.Var_name.t list

Takes in as args the Var_name.t name of the aliased object, the list of possible aliases, a region filter function that filters out regions needing linear capabilities, a list of the current live aliases and the block expression.

Returns the updated block expression along with the live variables at the start of the block.