Module Data_race_checker__Type_borrowing

val type_function_forward_borrowing_expr : Desugaring.Desugared_ast.class_defn list -> Desugaring.Desugared_ast.function_defn list -> Desugaring.Desugared_ast.expr -> unit Core.Or_error.t

check an identifier isn't borrowed multiple times when passed as args to function calls (forward borrowing)

val type_function_forward_borrowing_block_expr : Desugaring.Desugared_ast.class_defn list -> Desugaring.Desugared_ast.function_defn list -> Desugaring.Desugared_ast.block_expr -> unit Core.Or_error.t

val type_function_reverse_borrowing : Desugaring.Desugared_ast.class_defn list -> string -> Ast.Ast_types.type_expr -> Ast.Ast_types.borrowed_ref option -> Desugaring.Desugared_ast.block_expr -> unit Core.Or_error.t

Checks the function return type type_expr to determine if linear reference, whether it returns borrowed refernce borrowed_ref option and if the block_expr body expression returns a non-consumed id (not allowed if no reverse borrowing in function signature). string is the error prefix

val type_assign_borrowed_expr : Desugaring.Desugared_ast.class_defn list -> Desugaring.Desugared_ast.function_defn list -> Desugaring.Desugared_ast.expr -> Ast.Ast_types.borrowed_ref option Core.Or_error.t

Checks that borrowed values aren't assigned to objects (so don't outlive their borrowed scope.) Returns whether the current expr / block reduces to a borrowed value.

val type_assign_borrowed_block_expr : Desugaring.Desugared_ast.class_defn list -> Desugaring.Desugared_ast.function_defn list -> Desugaring.Desugared_ast.block_expr -> Ast.Ast_types.borrowed_ref option Core.Or_error.t