Module Data_race_checker__Type_consume_expr

Type checks the consume operation - ensures you can't consume or access already consumed variables

val type_consume_expr : Desugaring.Desugared_ast.class_defn list -> Desugaring.Desugared_ast.expr -> Desugaring.Desugared_ast.identifier list -> Desugaring.Desugared_ast.identifier list Core.Or_error.t

Takes as arguments the list of consumed variables before this expression, and returns the list of consumed variables after this expression is abstractly interpreted.

val type_consume_block_expr : Desugaring.Desugared_ast.class_defn list -> Desugaring.Desugared_ast.block_expr -> Desugaring.Desugared_ast.identifier list -> Desugaring.Desugared_ast.identifier list Core.Or_error.t