Module Data_race_checker__Type_linear_capabilities

This module type-checks potential accesses to linear capabilities of objects

val type_linear_obj_method_args : Desugaring.Desugared_ast.class_defn list -> Ast.Ast_types.Var_name.t -> Ast.Ast_types.Class_name.t -> Desugaring.Desugared_ast.identifier list -> Ast.Ast_types.loc -> unit Core.Or_error.t

Checks that linear objects are not passed as arguments to their methods.

val type_linear_args : Desugaring.Desugared_ast.class_defn list -> Desugaring.Desugared_ast.identifier list -> Ast.Ast_types.loc -> unit Core.Or_error.t

Checks that linear arguments are not duplicated in function calls.

val type_linear_capabilities_block_expr : Desugaring.Desugared_ast.class_defn list -> Desugaring.Desugared_ast.block_expr -> Desugaring.Desugared_ast.block_expr Core.Or_error.t