Module Data_race_checker__Type_region_annotations

This module checks that any region annotations to function/method are valid regions

val type_params_region_annotations : Desugaring.Desugared_ast.class_defn list -> Ast.Ast_types.param list -> unit Core.Or_error.t

This checks if the guards specified in function/method parameters are actually valid regions

val type_field_region_annotations : Ast.Ast_types.Class_name.t -> Ast.Ast_types.region list -> Ast.Ast_types.Region_name.t list -> Ast.Ast_types.region list Core.Or_error.t

This checks if the regions specified in fields are actually valid regions of the class